# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Norite REST API
  version: 0.0.0-m4
  description: >
    Source-of-truth REST contract for the Norite backend (see docs/architecture.md §2 and §10).
    `oapi-codegen` starts generating Go server-side types and routing scaffolding from this file at
    Milestone M12; until then it is hand-maintained, but the same rule already applies — every REST
    endpoint added or changed updates this file in the same commit (CLAUDE.md rule 6).


    IDs are Snowflakes (docs/adr/0003-snowflake-ids.md) and are always sent as **quoted strings**, never
    JSON numbers: they exceed 2^53, so a client parsing them as numbers silently loses precision and
    starts confusing adjacent IDs.
servers:
  - url: /api/v1

# Every operation requires a Bearer credential unless it declares `security: []`. Stating it this way round
# means a new endpoint is protected by default, and making one public is a visible, deliberate line in the
# diff rather than an omission nobody notices.
security:
  - bearerAuth: []
...
paths:
...
  /auth/refresh:
    post:
      summary: Rotate a refresh token
      description: >
        Exchanges a refresh token for a fresh pair, within the same device's family.


        Rotation is **single-use**: the presented token is revoked in the same transaction that creates its
        successor. Presenting it again is therefore unambiguous evidence of replay — either a client retried
        after a lost response, or the token was stolen. The two are indistinguishable from the server, so the
        pessimistic reading wins and **that device's entire session family is revoked**. Other devices are
        never affected.


        A token that was revoked *without* being rotated — by logout, or by a later login superseding it —
        is simply invalid and triggers none of that. Conflating the two would mean logging out, logging back
        in, and retrying the stale token once would revoke the session just created.


        Replay and an unknown token return the identical 401. Telling a caller "that token was already used"
        would confirm to a thief that they hold something real, and that the legitimate client got there
        first.
      operationId: refresh
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: A new token pair was issued; the presented refresh token is now spent.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPair"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401":
          description: >
            The refresh token is invalid, expired, revoked, or was already used. Replay additionally revokes
            that device's remaining sessions, but the response is the same either way.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429": { $ref: "#/components/responses/RateLimited" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >
        `Authorization: Bearer <credential>`, where the credential is one of two kinds and the server tells
        them apart by prefix:


        A **JWT access token** from `/auth/login` or `/auth/refresh` — HS256, 15-minute TTL, verified
        without a database round trip. It represents the person, so it passes every scope check
        (docs/adr/0022-access-token-signing-and-scope-model.md).


        An **API token** (`nat_…`) from `/auth/tokens` — long-lived, restricted to the scopes it was granted,
        and revocable instantly.


        Refresh tokens (`nrt_…`) are **not** Bearer credentials. They go in the body of `/auth/refresh` and
        `/auth/logout` and authenticate nothing else.

    operatorToken:
      type: http
      scheme: bearer
      description: >
        `Authorization: Bearer <operator token>` — the credential that authorizes instance administration
        before any account exists, and the only one in this system minted by a client rather than by the
        server.


        It is an HS256 JWT signed with the instance's own signing key (`[auth].jwt_secret`, or
        `NORITE_JWT_SECRET`), carrying `iss: "norite"`, `typ: "operator"`, an `exp`, and **no subject** — an
        operator is not an account. A token carrying a subject is refused, as is one with any other `typ`:
        without that check a device entry token would qualify, since it also names no account and carries
        the same issuer and key.


        What it proves is possession of the instance's configuration file, which is why it is the right
        authority for creating the first administrator and the wrong one for anything an account can do. It
        is not revocable — there is no row to delete — so it lives **two minutes** and is minted per
        request. `norite instance bootstrap` builds one; the shared definition is Go package
...
  schemas:
...
    TokenPair:
      type: object
      required: [access_token, refresh_token, token_type, expires_at]
      additionalProperties: false
      properties:
        access_token:
          type: string
          description: >
            A JWT, presented as `Authorization: Bearer …`. Short-lived by design: revocation acts on the
            refresh session, so a stolen access token stays usable until it expires and the TTL is what
            bounds that window.
        refresh_token:
          type: string
          description: >
            An opaque `nrt_…` value, stored server-side only as a SHA-256 hash. Send it to `/auth/refresh`
            or `/auth/logout`; it is never a Bearer credential. Single-use — each refresh returns a new one,
            and presenting a spent one revokes the device's whole family.
          examples: ["nrt_kZ8m1H0dQ2s7pXv4b9RtLcYwJfN6aE3gUiOoPqSdTbA"]
        token_type:
          type: string
          enum: [Bearer]
        expires_at:
          type: string
          format: date-time
          description: >
            When the **access** token expires, not the refresh token. Clients refresh on this.
...
    RefreshRequest:
      type: object
      required: [refresh_token]
      additionalProperties: false
      properties:
        refresh_token:
          type: string
...
    Error:
      description: >
        The envelope every non-2xx JSON response uses. Clients key off `code`, which is stable, never off
        `message`, which is human-readable and may change.
      type: object
      required: [error]
      additionalProperties: false
      properties:
        error:
          type: object
          required: [code, message]
          additionalProperties: false
          properties:
            code:
              type: string
              description: Stable, machine-readable error identifier.
              enum:
                - bad_request
                - unauthorized
                - forbidden
                - not_found
                - method_not_allowed
                - conflict
                - rate_limited
                - service_unavailable
                - internal_error
                # Still emitted by the OAuth sign-up path, which refuses outright on a gated instance —
                # there is nowhere to carry an invite code through a provider redirect. Registration's own
                # use of it ended at M10.
                - registration_closed
                - invite_required
                - invite_invalid
                - already_bootstrapped
                # The device-code grant's own vocabulary (RFC 8628 §3.5), from /auth/device/token. A
                # polling client branches on the first two every few seconds, so they are the codes here
                # that a generated client most needs to be able to name — and they were the ones missing
                # longest. Added after a manual run compared real responses against this enum; M9 shipped
                # them without this entry and nothing noticed, because the contract test walks the route
                # table and never a payload.
                - authorization_pending
                - slow_down
                - access_denied
                - expired_token
                # Configuration states, not faults. Each says a feature is switched off on this instance
                # rather than broken: no public base URL for the device flow, no mail relay for reset.
                - device_flow_unavailable
                - reset_unavailable
                # The OAuth paths, from M6.
                - oauth_email_unverified
                - oauth_provider_error
            message:
              type: string
              description: >
                Human-readable summary, safe to show a user.


                A 5xx message is generic — *except* where a handler has explicitly marked one as safe to
                send, which today is only the two `503`s above. Those say which feature is unconfigured,
                because "internal server error" would report an instance that is merely missing a setting
                as a broken one. What is never returned at any status is the underlying error: driver
                text, query fragments and hostnames are logged and stay there.
            request_id:
              type: string
              description: >
                Correlation ID for this request, matching the `X-Request-Id` response header and the
                server-side log line.
  responses:
    BadRequest:
      description: >
        The request body was malformed, exceeded 1 MiB, contained an unknown field, or failed validation.
        Unknown fields are rejected rather than ignored so a typo fails loudly instead of silently leaving
        a setting on its default (docs/architecture.md §14.1). Validation messages name the offending field
        and the rule it broke, never the submitted value — that value is a password on two of these routes.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    RateLimited:
      description: >
        Too many requests. All IP-based limiting groups IPv6 traffic by /64 subnet rather than exact
        address (docs/architecture.md §11), so every address within one /64 shares a single budget.


        The unauthenticated `/auth/*` routes additionally sit behind a stricter bucket than the rest of the
        API, counted independently: they are credential-guessing surfaces, and throttling a stuffing run
        must not also throttle a legitimate client's ordinary requests.
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds until the limit resets. Always present on a 429, and never `0`.
        X-RateLimit-Limit:
          schema:
            type: integer
        X-RateLimit-Remaining:
          schema:
            type: integer
        X-RateLimit-Reset:
          schema:
            type: integer
          description: Unix timestamp at which the current window resets.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"